requirements.txt
streamlit>=1.35.0
gspread>=6.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
protobuf>=3.20,<7
